Agregado pedido de reporte monopatines en Mantenimiento

+++ microservicioMonopatin/src/main/java/org/app/monopatin/entities/DTO/ReporteKmsDTO.java
@@ -3,6 +3,6 @@ package org.app.monopatin.entities.DTO;
 import org.app.monopatin.entities.Monopatin;
 
 public interface ReporteKmsDTO {
-    Integer getIdMonopatin();
+    Long getIdMonopatin();
     Double getKmsTotales();
 }
